packages/python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drtrace"
version = "0.1.0"
description = "Local log analysis client and daemon for Python applications"
requires-python = ">=3.8"
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "Thanh" }
]
dependencies = [
  "fastapi>=0.115,<0.116",
  "uvicorn[standard]>=0.30,<0.31",
  "psycopg2-binary>=2.9,<3.0",
  "pydantic>=2.8,<2.9"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "pytest-asyncio>=0.23,<0.24",
  "httpx>=0.27,<0.28"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
drtrace_service = [
    "resources/agents/*.md",
    "resources/agents/integration-guides/*.md",
    "resources/cpp/*.hpp"
]
drtrace = []  # Alias package - no additional data needed



